requirements.txt
asgiref==3.3.1
certifi==2020.11.8
cffi==1.14.3
cloudinary==1.23.0
cryptography==3.2.1
Django==3.1.3
django-rest-knox==4.1.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pytz==2020.4
simplejwt==2.0.1
six==1.15.0
sqlparse==0.4.1
typing==3.7.4.3
urllib3==1.26.2
whitenoise==5.2.0
